chore: first pass at consolidating javascript code
