--- src/theme/icons/CaretRight.js
@@ -1,30 +1,0 @@
-import PropTypes from 'prop-types';
-import Svg, { Polyline } from 'react-native-svg';
-
-import { makeIcon } from '@apollosproject/ui-kit';
-
-const Icon = makeIcon(({ size = 28, fill } = {}) => (
-  <Svg
-    width={size}
-    height={size}
-    viewBox="0 0 256 256"
-    fill={fill}
-    xmlns="http://www.w3.org/2000/Svg"
-  >
-    <Polyline
-      points="96 48 176 128 96 208"
-      fill="none"
-      stroke={fill}
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="16"
-    />
-  </Svg>
-));
-
-Icon.propTypes = {
-  size: PropTypes.number,
-  fill: PropTypes.string,
-};
-
-export default Icon;
