--- src/theme/icons/HandsHeart.js
@@ -1,28 +1,0 @@
-import PropTypes from 'prop-types';
-import Svg, { Path } from 'react-native-svg';
-
-import { makeIcon } from '@apollosproject/ui-kit';
-
-const Icon = makeIcon(({ size = 28, fill } = {}) => (
-  <Svg
-    width={size}
-    height={size}
-    viewBox="0 0 25 24"
-    fill={fill}
-    xmlns="http://www.w3.org/2000/Svg"
-  >
-    <Path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M16.249 1.84362L16.7688 1.32455C17.7986 0.296232 19.2575 -0.174065 20.6912 0.058304C22.1268 0.290966 23.3651 1.20023 24.0183 2.50501C24.8955 4.25786 24.551 6.37601 23.1773 7.74759L16.7855 14.4082C16.4914 14.7147 16.0078 14.7148 15.7136 14.4083L9.33282 7.76089C7.94866 6.37756 7.60433 4.26008 8.48081 2.50686C9.13399 1.20223 10.3723 0.292964 11.8079 0.0603012C13.2416 -0.172068 14.7005 0.298229 15.7298 1.32604L16.249 1.84362ZM22.1166 6.66607C23.0339 5.75005 23.2606 4.35603 22.6824 3.20054C22.2483 2.33348 21.4202 1.72541 20.4558 1.56912C19.4897 1.41253 18.5077 1.72909 17.8182 2.4176L16.7742 3.46013C16.4829 3.75109 16.017 3.75132 15.7253 3.46065L14.6809 2.4196C13.9914 1.73108 13.0094 1.41453 12.0433 1.57111C11.0789 1.72741 10.2508 2.33548 9.81689 3.20213C9.23904 4.358 9.46571 5.75189 10.3936 6.6795L16.2495 12.7798L22.1166 6.66607ZM0.5 15.7456V14.2477C0.5 13.8253 0.835777 13.4828 1.24998 13.4828C1.66418 13.4828 1.99996 13.8253 1.99996 14.2477V14.9807H6.49982C8.30681 14.9807 9.81961 16.2641 10.1722 17.9765H13.2496C15.3171 17.9765 16.9995 19.6566 16.9995 21.7372C16.9995 22.1596 16.6637 22.5021 16.2495 22.5021H1.99996V23.2351C1.99996 23.6575 1.66418 24 1.24998 24C0.835777 24 0.5 23.6575 0.5 23.2351V21.7372V15.7456ZM1.99996 16.5105V20.9723H15.3637C15.0493 20.1185 14.2234 19.5063 13.2496 19.5063H9.49974H6.49982C6.08562 19.5063 5.74985 19.1639 5.74985 18.7414C5.74985 18.319 6.08562 17.9765 6.49982 17.9765H8.61389C8.29945 17.1227 7.47361 16.5105 6.49982 16.5105H1.99996Z"
-      fill={fill}
-    />
-  </Svg>
-));
-
-Icon.propTypes = {
-  size: PropTypes.number,
-  fill: PropTypes.string,
-};
-
-export default Icon;
