--- src/theme/icons/NextSteps.js
@@ -1,38 +1,0 @@
-import PropTypes from 'prop-types';
-import Svg, { Polyline } from 'react-native-svg';
-
-import { makeIcon } from '@apollosproject/ui-kit';
-
-const Icon = makeIcon(({ size = 32, fill } = {}) => (
-  <Svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={size}
-    height={size}
-    fill={fill}
-    viewBox="0 0 256 256"
-  >
-    <Polyline
-      points="232.002 56 136.002 152 96.002 112 24.002 184"
-      fill="none"
-      stroke={fill}
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="16"
-    />
-    <Polyline
-      points="232.002 120 232.002 56 168.002 56"
-      fill="none"
-      stroke={fill}
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="16"
-    />
-  </Svg>
-));
-
-Icon.propTypes = {
-  size: PropTypes.number,
-  fill: PropTypes.string,
-};
-
-export default Icon;
