--- src/theme/icons/index.js
@@ -1,10 +1,0 @@
-/* Export your custom icon components here. */
-export Badge from './Badge';
-export CaretRight from './CaretRight';
-export HandsHeart from './HandsHeart';
-export HandWaving from './HandWaving';
-export House from './House';
-export LetterE from './LetterE';
-export NextSteps from './NextSteps';
-export Question from './Question';
-export WatchVideo from './WatchVideo';
