--- src/theme/index.js
@@ -1,4 +1,0 @@
-import theme from './theme';
-import * as customIcons from './icons';
-
-export { theme as default, customIcons };
