--- src/theme/theme.js
@@ -1,127 +1,0 @@
-import { Text } from 'react-native';
-import { ScriptureText, VerseNumber } from '@apollosproject/ui-scripture';
-
-/* Add your custom theme definitions below. Anything that is supported in UI-Kit Theme can be
- overridden and/or customized here! */
-
-/* Base colors.
- * These get used by theme types (see /types directory) to color
- * specific parts of the interface. For more control on how certain
- * elements are colored, go there. The next level of control comes
- * on a per-component basis with "overrides"
- */
-const colors = {
-  primary: '#952124',
-  secondary: '#9A3434',
-  tertiary: '#8E8E93',
-};
-
-/* Base Typography sizing and fonts.
- * To control speicfic styles used on different type components (like H1, H2, etc), see "overrides"
- */
-const typography = {
-  sans: {
-    regular: {
-      default: 'Futura',
-      italic: 'Futura-Italic',
-    },
-    medium: {
-      default: 'Futura-Medium',
-      italic: 'Futura-MediumItalic',
-    },
-    bold: {
-      default: 'Futura-Bold',
-      italic: 'Futura-BoldItalic',
-    },
-    black: {
-      default: 'Futura-Bold',
-      italic: 'Futura-BoldItalic',
-    },
-  },
-  ui: {
-    regular: 'System',
-  },
-};
-
-/* Responsive breakpoints */
-// export const breakpoints = {};
-
-/* Base sizing units. These are used to scale
- * space, and size components relatively to one another.
- */
-// export const sizing = {};
-
-/* Base alpha values. These are used to keep transparent values across the app consistant */
-// export const alpha = {};
-
-/* Base overlays. These are used as configuration for LinearGradients across the app */
-// export const overlays = () => ({});
-
-/* Overrides allow you to override the styles of any component styled using the `styled` HOC. You
- * can also override the props of any component using the `withTheme` HOC. See examples below:
- * ```const StyledComponent = styled({ margin: 10, padding: 20 }, 'StyledComponent');
- *    const PropsComponent = withTheme(({ theme }) => ({ fill: theme.colors.primary }), 'PropsComponent');
- * ```
- * These componnents can have their styles/props overriden by including the following overrides:
- * ```{
- *   overides: {
- *     StyledComponent: {
- *       margin: 5,
- *       padding: 15,
- *     },
- *     // #protip: you even have access 👇to component props! This applies to style overrides too 💥
- *     PropsComponent: () => ({ theme, isActive }) => ({
- *       fill: isActive ? theme.colors.secondary : theme.colors.primary,
- *     }),
- *   },
- * }
- * ```
- */
-const overrides = {
-  'ui-onboarding.Landing.BrandIcon': {
-    name: 'letter-e', // Use this if we need to add a custom brand icon
-  },
-  'ui-onboarding.Landing': {
-    description:
-      'The Church of Eleven22 is a movement for all people to discover and deepen a relationship with Jesus Christ.',
-  },
-  'ui-onboarding.Landing.Subtitle': (theme) => ({
-    color: theme.colors.text.primary,
-  }),
-  'ui-onboarding.Features': {
-    description:
-      "We'd like to help you personalize your profile to make the most of your online experience.",
-  },
-  'ui-onboarding.LocationFinder': {
-    description:
-      "We'll use your location to connect you with your nearby campus and community.",
-  },
-  'ui-onboarding.AskNotifications': {
-    description:
-      "We'll let you know when important things are happening and keep you in the loop.",
-  },
-  'ui-onboarding.Follow': {
-    description: 'Follow others to stay up to date on your church community.',
-  },
-  'ui-onboarding.Entry': {
-    authTitleTest: 'Have We Met?',
-    description:
-      'Sign in for a personalized experience as you discover and deepen your relationship with Jesus Christ',
-  },
-  'ui-auth.ProfileEntry': {
-    prompt: 'Complete your profile to get connected!',
-  },
-  'ui-prayer.Confirmation': {
-    body: () => (
-      <Text>
-        <ScriptureText>
-          And this is the confidence that we have toward him, that if we ask
-          anything according to his will he hears us.{'\n'}
-          <VerseNumber>1 John 5:14</VerseNumber>
-        </ScriptureText>
-      </Text>
-    ),
-  },
-};
-
-export default { colors, overrides, typography };
